notifier: move notify_data definition to header
Moves notify_data struct definition to notifier header.
