+++ src/include/sof/lib/notifier.h
@@ -8,6 +8,8 @@
 #ifndef __SOF_LIB_NOTIFIER_H__
 #define __SOF_LIB_NOTIFIER_H__
 
+#include <sof/common.h>
+#include <sof/lib/memory.h>
 #include <sof/list.h>
 #include <sof/spinlock.h>
 #include <stdint.h>
@@ -34,6 +36,13 @@ struct notify {
 	spinlock_t *lock;	/* notifier lock */
 	struct list_item list[NOTIFIER_ID_COUNT]; /* list of callback handles */
 };
+
+struct notify_data {
+	void *caller;
+	enum notify_id type;
+	uint32_t data_size;
+	void *data;
+} __aligned(PLATFORM_DCACHE_ALIGN);
 
 #ifdef CLK_SSP
 #define NOTIFIER_CLK_CHANGE_ID(clk) \
